move scm-slang to dedicated alt-lang folder
