+++ .eslintignore
@@ -1,3 +1,3 @@
 /dist/
-/src/scm-slang/
+/src/alt-langs/
 /src/py-slang/
