+++ .prettierignore
@@ -1,3 +1,3 @@
-/src/scm-slang/
+/src/alt-langs/
 /src/py-slang/
 /src/**/__tests__/**/__snapshots__
